+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Using `goid` to log goroutine ID for debugging
+
 ## [0.4.4] - 2026-01-24
 
 ### Changed